--story=859166189 【scm/prpject】监控工蜂接口以及users接口的情况，上报monitor

+++ src/backend/ci/ext/tencent/monitoring/biz-monitoring-tencent/src/main/kotlin/com/tencent/devops/monitoring/dao/SlaDailyDao.kt
@@ -11,7 +11,7 @@ class SlaDailyDao {
         dslContext: DSLContext,
         module: String,
         name: String,
-        percent: Int,
+        percent: Double,
         startTime: LocalDateTime,
         endTime: LocalDateTime
     ) {
